refactor(testutil): split minio/rabbitmq testcontainer helpers into subpackages [#1908]
Move minio_integration.go and rabbitmq_integration.go from package testutil
into tests/testutil/minioctr and tests/testutil/rabbitmqctr respectively, so
that importing tests/testutil no longer drags testcontainers-go/modules/minio
and modules/rabbitmq into every consumer's module graph.
