+++ tests/testutil/minioctr/minioctr_test.go
@@ -1,6 +1,6 @@
 //go:build integration
 
-package testutil
+package minioctr
 
 import (
 	"context"
